feat: move db to folder; add user, settings db queries
