+++ package.json
@@ -24,6 +24,7 @@
         "vite": "3.2.0"
     },
     "dependencies": {
-        "@svelteuidev/core": "^0.15.3"
+        "@svelteuidev/core": "^0.15.3",
+        "uuid": "^9.0.0"
     }
 }
