+++ src/utils.ts
@@ -1,4 +1,6 @@
 export enum MessageRequest {
-    SAVE_TAB,
-    GET_ALL_ITEMS
+  SAVE_TAB,
+  GET_ALL_ITEMS,
+  GET_SETTINGS,
+  UPDATE_SETTINGS,
 }
